-- get rid of inheritance from AbstractConsoleRunner with history. It adds no almost value as we override most of the methods. On the other hand it makes the flow more complicated,  forces us to redeclare many fields , and forces casts (as it does not use generic types) . Removing the inheritance also allows to make most methods private --  Extract static utility methods in PydevConsoleRunner into an interface and move the implementation into PydevConsoleRunnerImpl
